Tests/LibWeb/Text/expected/Editing/execCommand-fontName.txt

foo<font face="monospace">bar</font>
<font face="sans-serif">foobar</font>
